Reset repository to match main Cwmfest repository
This repository was supposed to be where I worked on the Cwmfest site / app before pushing changes to the main Cwmfest repository, so that any mistakes I made while learning new techniques wouldn't affect the site or app. I ended up working directly on the main repository for a while, and ignoring this one, so I've wiped it and pushed the changes here to give me a fresh start for practising.